Solved "complication" when cleaning and rebuilding the virtualenv.
The Makefile target to create the virtual env (venv) simply called python,
e.g. python3.8, without a full path. If the virtualenv was active when
you called `make clean-venv venv` you'd be trying to create the virtualenv
using the version of python you'd just deleted. That didn't always end so
well, particularly if you were using a tool such as direnv as it left the
current vritualenv active.

To fix this the site (or system) python is used to create the virtualenv,
using env to make it a bit more portable.