Split the requirements into different files
With the requirements split into separate files, dev, docs and tests
it simplifies the setup.cfg and ensures that only the required packages
are installed for a given function. The same applies with third-party
services such as read the docs so we use the the free service repsonsibly
and don't download or install things that are not needed (ignore that the
packages are cached).

The downside is that pinning the package version numbers for the separate
files now becomes impossible without pip-compile so pip-tools was made
"compulsory". Since this is 2020 this is probably how it should be.