Skip creating the virtualenv and repository in development
The post-generation hook creates the virtualenv and intialises the
repository. In development, the project is generated multiple times
so the ability to skip these steps really speeds things up. It also
solves two problems:

1. When pip-compile generates the requirements/dev.txt it replaces
   the relative path when initialising the package for development,
   "-e ." with the absolute path on the local machine. That meant
   requirement/dev.txt had to be reverted every time the project
   was generated.

2. Initialising the git repository if it already existed did not
   appear to have any side-effects. However setting the origin if
   it was already set generated an error. Generally this was not
   a problem but it did not look good and strictly speaking the
   post-generation hook should report the error to trigger the
   cleanup process.

Since the project is generated by passing the repository URL to
cookiecutter, setting the environment variable that controls
whether the virtualenv or repository are skipped in a .env file
will not interfere with normal use of the template.