Remove whitespace in the Makefile target, make checks
Depending on which tools were selected the make check target could
have blank lines. If the line began with a <tab> then all the steps
would still be executed. However if the file was ever opened with
an editor it was likely that the trailing whitespace would be removed
causing the target to fail next time it was run.