Show that project keywords is a comma-separated list
The keywords argument to setup() is poorly documented. Some sources
report it being a list of strings, a list of keywords separated by
spaces or a list of keywords separated by commas. The confusion is
likely due to the existance of setup.cfg and the syntax used there.
Since we are entering keywords as a single string it seems likely
that the comma-separated list since that maps to the ini format
used in setup.cfg.