No app urls are defined if project is not created
When the example project is included in the generated project urls.py
contains the example view, IndexView. When the create_project parameter
was set to "n" (No) the view was still included. The reason was the
template if statement were checking simply for the parameter being
defined rather than it being set to "y".