Replace the todo list with a roadmap
The TODO.txt was always a work in progress. Since it was always under
revision not all the detailed were fleshed out and that made it hard
to understand what was being said. A roadmap is more stable. It is
still subject to change but that should be over a longer period of
time and so likely to be more informative.