Add an outline for a guide for contributors to the template.
Rather than add a standard form guide which is customized with the
name of the project when the project is generated a general outline
with comments was added. That gives the guide author some idea of
the content and so make it easier to write. it also means that the
final result will have the same content style as the project and
as a result be more genuine.