Change the output directory to 'output' within the project
The output directory was previously set to be the parent directory
so the generated project would be a peer of the template. That was
done as there generated project was also hosted on GitHub so you
could see what was generated and also deployed to PyPi to show that
the template worked out of the box.

However the result was that development and debugging was a little
less convenient so the output directory was moved to inside the
project tree so browsing files was easier.