Add support for direnv and other virtualenv managers
Reorganised the Makefile to allow the path to the bin directory where
pip, etc., lives to be overridden in an included makefile. That makes
it easy to use tools for managing virtualenvs, direnv, pyenv, etc.
and still be able to use the makefile - which relies on the full path
to a given tool rather than activating the virtualenv first.

Moved the the '-include' command from the end of the Makefile to the
beginning - just after the variable definitions so it can be used to
override those and add new targets.

direnv's .envrc file uses the standard python layout. The direnv
wiki, https://github.com/direnv/direnv/wiki/Python does have code
for using virtualenvs creating using the venv module but there's no
pressing need to use that here.

Separated out the commands to install the dependencies into a new
'install' target so it's easier to work with diremv, etc.