Set initial version to 0.0.0 so bump2version can managed it
Previously the initial version of the generated project was set as a
parameter when the project was generated. However unless the format
followed <major>.<minor>.<patch> bump2version was unable to parse it
correctly. Since most projects using the template will be new setting
the initial version to 0.0.0 means that bump2version can be used right
away to manage the release versions.