Simplify managing the project requirements
Updated the install target so if any of the input files change then
pip-compile is run to generate the corresponding .txt file and pip-sync
updates the virtual env. That was done by:

1. Add rules for installing pip and pip-tools.
2. Add a rule for creating the virtualenv
3. Add rules for building each and all of the requirements files.

Now `make install` can be used to create the virtualenv and install all
of the dependencies and update and sync the virtualenv if any new
packages are added later.