Fix the 'messages' target so it can be run from the root directory
The makemessages management command would collect the the translation
strings from the project file but the compilemessages management
command would fail to generate the .po file.

Note also that the virtualenv directory has to be specifically excluded
to avoid all the translation strings in the installed packges from being
compiled also.